refactor: migrate history, summary, analytics, and github routes to routers
