+++ api/api.py
@@ -7,7 +7,7 @@ import logging
 from fastapi import FastAPI, HTTPException
 from fastapi.responses import RedirectResponse
 
-from api.routers import summarise as summarise_router
+from api.routers import summarise as summarise_router, history, summary, analytics, github
 from fastapi.middleware.cors import CORSMiddleware
 from pydantic import BaseModel
 from typing import List, Optional
@@ -47,6 +47,10 @@ async def lifespan(app: FastAPI):
 
 app = FastAPI(title="gitpulse API", version="0.6.0", lifespan=lifespan)
 app.include_router(summarise_router.router)
+app.include_router(history.router)
+app.include_router(summary.router)
+app.include_router(analytics.router)
+app.include_router(github.router)
 
 # CORS
 app.add_middleware(
@@ -57,99 +61,7 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-# Models
-class SummariseRequest(BaseModel):
-    username: str
-    repos: List[str]
-    days: int = 7
-
-class SummariseResponse(BaseModel):
-    id: str
-    display: str
-    summary: str
-    repos: List[str]
-    days: int
-    username: str
-    generated_at: str
-    is_public: bool = False
-
-class RosterRequest(BaseModel):
-    name: str
-    usernames: List[str]
-
-class RosterResponse(BaseModel):
-    id: str
-    name: str
-    usernames: List[str]
-    created_at: str
-
-class TeamSummariseRequest(BaseModel):
-    usernames: List[str]
-    repos: List[str]
-    days: int = 7
-
-class TeamSummariseResponse(BaseModel):
-    display: str
-    summary: str
-    repos: List[str]
-    days: int
-    contributors: List[str]
-    generated_at: str
-
-class SlackDeliverRequest(BaseModel):
-    summary: str
-    webhook_url: str
-    channel: Optional[str] = None
-
-class PublicToggleRequest(BaseModel):
-    public: bool
-
-class PublicToggleResponse(BaseModel):
-    id: str
-    is_public: bool
-
-class PublicSummaryResponse(BaseModel):
-    id: str
-    username: str
-    repos: List[str]
-    days: int
-    summary: str
-    generated_at: str
-
-class CompareRecord(BaseModel):
-    commits: int
-    prs: int
-    issues: int
-    active_days: int
-
-class CompareResponse(BaseModel):
-    username: str
-    days: int
-    current: CompareRecord
-    previous: CompareRecord
-    delta: dict
-
-
-class RecommendationsRequest(BaseModel):
-    username: str
-    days: int = 30
-
-class RecommendationsResponse(BaseModel):
-    recommendations: str
-    generated_at: str
-
-class PromptTemplateCreate(BaseModel):
-    username: str
-    name: str
-    content: str
-
-class PromptTemplateResponse(BaseModel):
-    id: str
-    username: str
-    name: str
-    content: str
-    created_at: str
-
+from api.models import *
 
 # Routes
 @app.get("/health")
@@ -201,569 +113,6 @@ async def health_keys():
     return results
 
 
-
-@app.patch("/history/{summary_id}/public", response_model=PublicToggleResponse)
-async def toggle_summary_public(summary_id: str, req: PublicToggleRequest):
-    """Toggle the is_public flag for a summary."""
-    pool = get_db_pool()
-    if not pool:
-        raise HTTPException(status_code=503, detail="Database integration disabled")
-    try:
-        async with pool.acquire() as conn:
-            row = await conn.fetchrow(
-                "UPDATE summaries SET is_public = $1 WHERE id::text = $2 RETURNING id, is_public",
-                req.public, summary_id
-            )
-            if not row:
-                raise HTTPException(status_code=404, detail="Summary not found")
-            return PublicToggleResponse(id=str(row['id']), is_public=row['is_public'])
-    except HTTPException: raise
-    except Exception as e:
-        logger.error("Failed to toggle public status: %s", e)
-        raise HTTPException(status_code=500, detail="Failed to toggle public status")
-
-@app.get("/summary/public/{summary_id}", response_model=PublicSummaryResponse)
-async def get_public_summary(summary_id: str):
-    """Fetch a public summary by ID without authentication."""
-    pool = get_db_pool()
-    if not pool:
-        raise HTTPException(status_code=503, detail="Database integration disabled")
-    try:
-        async with pool.acquire() as conn:
-            row = await conn.fetchrow(
-                "SELECT id, username, repos, days, summary, generated_at FROM summaries WHERE id::text = $1 AND is_public = TRUE",
-                summary_id
-            )
-            if not row:
-                raise HTTPException(status_code=404, detail="Public summary not found")
-            return PublicSummaryResponse(
-                id=str(row['id']),
-                username=row['username'],
-                repos=row['repos'],
-                days=row['days'],
-                summary=row['summary'],
-                generated_at=row['generated_at'].strftime("%Y-%m-%dT%H:%M:%SZ")
-            )
-    except HTTPException: raise
-    except Exception as e:
-        logger.error("Failed to fetch public summary: %s", e)
-        raise HTTPException(status_code=500, detail="Failed to fetch public summary")
-
-@app.get("/analytics/compare", response_model=CompareResponse)
-async def compare_periods(username: str, days: int = 30):
-    """Compare performance metrics between current and previous periods."""
-    repos = await _get_user_repos(username)
-    if not repos:
-        raise HTTPException(status_code=404, detail="No repositories found for user")
-
-    # Current period
-    activity_curr, _ = await get_activity(source="github", username=username, repos=repos, days=days)
-    
-    # Previous period
-    # To get the previous period of the same length, we look back from (days) ago to (2*days) ago.
-    # However, get_activity is built to look back N days from NOW.
-    # We might need a more flexible get_activity or a custom implementation here.
-    # For now, let's assume we can use a custom fetch since get_activity is a bit fixed.
-    
-    # Actually, let's just do two calls if we can but handle the dates.
-    # current: [now - days, now]
-    # previous: [now - 2*days, now - days]
-    
-    # I'll implement a helper for this or use raw calls.
-    # Given the complexity, I'll stick to a slightly simplified comparison for this sprint:
-    # We'll fetch all activity for 2*days and split it.
-    
-    full_activity, _ = await get_activity(source="github", username=username, repos=repos, days=days * 2)
-    
-    now = datetime.now(timezone.utc)
-    split_date = now - timedelta(days=days)
-    
-    def process_activity(items, date_key, is_current):
-        count = 0
-        active_dates = set()
-        for item in items:
-            dt = item.get(date_key)
-            if dt:
-                if is_current and dt >= split_date:
-                    count += 1
-                    active_dates.add(dt.date())
-                elif not is_current and dt < split_date:
-                    count += 1
-                    active_dates.add(dt.date())
-        return count, len(active_dates)
-
-    # Process Current
-    curr_commits, curr_active_commit_days = process_activity(full_activity.get("commits", []), "date", True)
-    curr_prs, curr_active_pr_days = process_activity(full_activity.get("prs", []), "merged_at", True)
-    curr_issues, curr_active_issue_days = process_activity(full_activity.get("issues", []), "closed_at", True)
-    curr_active_days = len(set(list(full_activity.get("commits", [])) + list(full_activity.get("prs", [])) + list(full_activity.get("issues", [])))) # Simplified
-    
-    # Re-calculate active days properly
-    curr_active_total = set()
-    prev_active_total = set()
-    
-    for c in full_activity.get("commits", []):
-        d = c["date"]
-        if d >= split_date: curr_active_total.add(d.date())
-        else: prev_active_total.add(d.date())
-    for p in full_activity.get("prs", []):
-        d = p["merged_at"]
-        if d >= split_date: curr_active_total.add(d.date())
-        else: prev_active_total.add(d.date())
-    for i in full_activity.get("issues", []):
-        d = i["closed_at"]
-        if d >= split_date: curr_active_total.add(d.date())
-        else: prev_active_total.add(d.date())
-
-    # Process Previous
-    prev_commits, _ = process_activity(full_activity.get("commits", []), "date", False)
-    prev_prs, _ = process_activity(full_activity.get("prs", []), "merged_at", False)
-    prev_issues, _ = process_activity(full_activity.get("issues", []), "closed_at", False)
-
-    current = CompareRecord(commits=curr_commits, prs=curr_prs, issues=curr_issues, active_days=len(curr_active_total))
-    previous = CompareRecord(commits=prev_commits, prs=prev_prs, issues=prev_issues, active_days=len(prev_active_total))
-
-    def calc_delta(curr, prev):
-        if prev == 0: return 100 if curr > 0 else 0
-        return round(((curr - prev) / prev) * 100, 1)
-
-    delta = {
-        "commits": calc_delta(curr_commits, prev_commits),
-        "prs": calc_delta(curr_prs, prev_prs),
-        "issues": calc_delta(curr_issues, prev_issues),
-        "active_days": calc_delta(len(curr_active_total), len(prev_active_total))
-    }
-
-    return CompareResponse(username=username, days=days, current=current, previous=previous, delta=delta)
-
-
-@app.get("/history")
-async def get_history(
-    username: str, 
-    limit: int = 10, 
-    search: str | None = None,
-    start_date: str | None = None,
-    end_date: str | None = None
-):
-    """
-    Fetch historical summaries for a given username with filtering options.
-    """
-    logger.info("Fetching history for %s (limit: %d, search: %s, date: %s-%s)", username, limit, search, start_date, end_date)
-    # Validate dates early
-    start_dt = None
-    end_dt = None
-    try:
-        if start_date:
-            start_dt = datetime.strptime(start_date, "%Y-%m-%d")
-        if end_date:
-            end_dt = datetime.strptime(end_date, "%Y-%m-%d")
-    except ValueError:
-        raise HTTPException(status_code=400, detail="Invalid date format. Use YYYY-MM-DD")
-
-    pool = get_db_pool()
-    if not pool:
-        logger.warning("DB pool not initialized. Cannot fetch history.")
-        return {"summaries": [], "total": 0}
-
-    try:
-        async with pool.acquire() as connection:
-            query = """
-                SELECT id, username, repos, days, summary, generated_at
-                FROM summaries
-                WHERE username = $1
-            """
-            params = [username]
-            idx = 2
-            
-            if search:
-                query += f" AND (repos::text ILIKE ${idx} OR summary ILIKE ${idx})"
-                params.append(f"%{search}%")
-                idx += 1
-                
-            if start_dt:
-                query += f" AND generated_at >= ${idx}"
-                params.append(start_dt)
-                idx += 1
-                
-            if end_dt:
-                query += f" AND generated_at <= ${idx}"
-                params.append(end_dt)
-                idx += 1
-                
-            query += f" ORDER BY generated_at DESC LIMIT ${idx}"
-            params.append(limit)
-            
-            records = await connection.fetch(query, *params)
-            
-            # Count also needs filters
-            count_query = "SELECT COUNT(*) FROM summaries WHERE username = $1"
-            count_params = [username]
-            c_idx = 2
-            if search:
-                count_query += f" AND (repos::text ILIKE ${c_idx} OR summary ILIKE ${c_idx})"
-                count_params.append(f"%{search}%")
-                c_idx += 1
-            if start_dt:
-                count_query += f" AND generated_at >= ${c_idx}"
-                count_params.append(start_dt)
-                c_idx += 1
-            if end_dt:
-                count_query += f" AND generated_at <= ${c_idx}"
-                count_params.append(end_dt)
-            
-            total_count = await connection.fetchval(count_query, *count_params) or 0
-            
-            return {
-                "summaries": [
-                    {
-                        "id": str(r["id"]),
-                        "username": r["username"],
-                        "repos": r["repos"],
-                        "days": r["days"],
-                        "summary": r["summary"],
-                        "generated_at": r["generated_at"].strftime("%Y-%m-%dT%H:%M:%SZ")
-                    }
-                    for r in records
-                ],
-                "total": total_count or 0
-            }
-    except Exception as e:
-        logger.error("msg: %s", e, exc_info=True)
-        raise HTTPException(status_code=500, detail={"error": "Failed to fetch history.", "code": 500})
-
-async def _get_user_repos(username: str) -> list[str]:
-    # Check cache first (10 minute expiry)
-    repos = repo_cache.get(username)
-    if repos:
-        logger.info("Using cached repo list for %s", username)
-        return repos
-
-    headers = {
-        "Accept": "application/vnd.github+json",
-        "X-GitHub-Api-Version": "2022-11-28"
-    }
-    token = os.getenv("GITHUB_TOKEN")
-    if token:
-        headers["Authorization"] = f"Bearer {token}"
-        
-    url = f"https://api.github.com/users/{username}/repos"
-    
-    async with httpx.AsyncClient() as client:
-        params = {"type": "public", "per_page": 100}
-        try:
-            response = await client.get(url, headers=headers, params=params, timeout=15.0)
-            if response.status_code == 401:
-                logger.error("GitHub API 401 Unauthorized for %s. Check GITHUB_TOKEN.", username)
-                raise HTTPException(status_code=401, detail="GitHub Token is invalid or expired. Please check your .env file.")
-            response.raise_for_status()
-            data = response.json()
-            repos = [repo["name"] for repo in data]
-            
-            # Update cache
-            repo_cache.set(username, repos, ttl=600) # 10 min for repo list
-            return repos
-        except Exception as e:
-            logger.error("Failed to fetch repos for %s: %s", username, e)
-            # If fetch fails but we have some cache, return it as fallback (even if expired)
-            cached_repos = repo_cache._cache.get(username) # Access internal for fallback
-            if cached_repos:
-                logger.warning("Returning stale/expired repo list for %s as fallback", username)
-                return cached_repos[0]
-            raise HTTPException(status_code=500, detail="Failed to fetch user repositories from GitHub")
-
-@app.get("/analytics/commits-per-day")
-async def get_activity_per_day(username: str, days: int = 30):
-    repos = await _get_user_repos(username)
-    if not repos:
-        return []
-    
-    try:
-        activity, errors = await get_activity(source="github", username=username, repos=repos, days=days)
-        commits = activity.get("commits", [])
-    except Exception as e:
-        logger.error("Failed to fetch commits: %s", e)
-        raise HTTPException(status_code=500, detail="Failed to fetch commits")
-        
-    counts = Counter()
-    for commit in commits:
-        date_str = commit["date"].strftime("%Y-%m-%d")
-        counts[date_str] += 1
-        
-    result = [{"date": k, "count": v} for k, v in sorted(counts.items())]
-    return result
-
-@app.get("/analytics/repos-breakdown")
-async def get_repos_breakdown(username: str, days: int = 30):
-    repos = await _get_user_repos(username)
-    if not repos:
-        return []
-        
-    try:
-        activity, errors = await get_activity(source="github", username=username, repos=repos, days=days)
-        commits = activity.get("commits", [])
-    except Exception as e:
-        logger.error("Failed to fetch commits: %s", e)
-        raise HTTPException(status_code=500, detail="Failed to fetch commits")
-        
-    counts = Counter()
-    for commit in commits:
-        counts[commit["repo"]] += 1
-        
-    total_commits = sum(counts.values())
-    
-    result = []
-    if total_commits > 0:
-        for repo, count in counts.items():
-            result.append({
-                "repo": repo,
-                "count": count,
-                "percentage": round((count / total_commits) * 100, 1)
-            })
-    return sorted(result, key=lambda x: x["count"], reverse=True)
-
-@app.get("/github/validate")
-async def validate_github_user(username: str):
-    """
-    Check if a GitHub user exists and return their profile info.
-    """
-    logger.info("Validating GitHub user: %s", username)
-    headers = {
-        "Accept": "application/vnd.github+json",
-        "X-GitHub-Api-Version": "2022-11-28"
-    }
-    token = os.getenv("GITHUB_TOKEN")
-    if token:
-        headers["Authorization"] = f"Bearer {token}"
-        
-    async with httpx.AsyncClient() as client:
-        try:
-            response = await client.get(f"https://api.github.com/users/{username}", headers=headers, timeout=10.0)
-            if response.status_code == 200:
-                data = response.json()
-                # Also fetch repos to populate cache and return count
-                repos = await _get_user_repos(username)
-                return {
-                    "valid": True, 
-                    "username": data["login"], 
-                    "avatar_url": data["avatar_url"],
-                    "repos": repos
-                }
-            elif response.status_code == 404:
-                return {"valid": False, "error": "User not found"}
-            else:
-                return {"valid": False, "error": f"GitHub API error: {response.status_code}"}
-        except Exception as e:
-            logger.error("Error validating user %s: %s", username, e)
-            return {"valid": False, "error": str(e)}
-
-@app.get("/github/repos")
-async def get_github_repos(username: str):
-    """
-    Fetch list of repos for a user (with caching).
-    """
-    try:
-        repos = await _get_user_repos(username)
-        return {"repos": repos}
-    except HTTPException as e:
-        raise e
-    except Exception as e:
-        logger.error("Error fetching repos for %s: %s", username, e)
-        raise HTTPException(status_code=500, detail="Failed to fetch repositories")
-
-@app.get("/analytics/all")
-async def get_analytics_full(username: str, days: int = 30, refresh: bool = False):
-    """
-    Fetch all analytics data in a single optimized pass.
-    """
-    # Check cache
-    cache_key = f"analytics:{username}:{days}"
-    if not refresh:
-        cached = analytics_cache.get(cache_key)
-        if cached:
-            logger.info("Using cached analytics for %s", username)
-            return cached
-
-    # 1. Get total summaries from DB
-    pool = get_db_pool()
-    total_summaries = 0
-    if pool:
-        try:
-            async with pool.acquire() as conn:
-                count = await conn.fetchval('SELECT COUNT(*) FROM summaries')
-                total_summaries = count or 0
-        except Exception as e:
-            logger.error("DB error figuring out summaries count: %s", e)
-            
-    # 2. Get user repos
-    repos = await _get_user_repos(username)
-    if not repos:
-        return {
-            "commits_per_day": [],
-            "repos_breakdown": [],
-            "insights": {
-                "most_active_day": "N/A", "streak": 0, "top_repo": "N/A",
-                "total_summaries": total_summaries, "average_commits_per_day": 0
-            }
-        }
-        
-    # 3. Get ALL relevant commits in ONE sweep
-    try:
-        activity, errors = await get_activity(source="github", username=username, repos=repos, days=days)
-        commits = activity.get("commits", [])
-    except Exception as e:
-        logger.error("Failed to fetch commits for %s: %s", username, e)
-        # Return empty data instead of 500 to keep dashboard stable
-        return {
-            "commits_per_day": [],
-            "repos_breakdown": [],
-            "insights": {
-                "most_active_day": "N/A", "streak": 0, "top_repo": "N/A",
-                "total_summaries": total_summaries, "average_commits_per_day": 0
-            }
-        }
-        
-    if not commits:
-         return {
-            "commits_per_day": [],
-            "repos_breakdown": [],
-            "insights": {
-                "most_active_day": "N/A", "streak": 0, "top_repo": "N/A",
-                "total_summaries": total_summaries, "average_commits_per_day": 0
-            }
-        }
-
-    # 4. Process data (Frequency)
-    counts_freq = Counter()
-    for commit in commits:
-        date_str = commit["date"].strftime("%Y-%m-%d")
-        counts_freq[date_str] += 1
-    commits_per_day = [{"date": k, "count": v} for k, v in sorted(counts_freq.items())]
-
-    # 5. Process data (Breakdown)
-    counts_repo = Counter()
-    for commit in commits:
-        counts_repo[commit["repo"]] += 1
-    
-    total_commits = len(commits)
-    repos_breakdown = []
-    for repo, count in counts_repo.items():
-        repos_breakdown.append({
-            "repo": repo,
-            "count": count,
-            "percentage": round((count / total_commits) * 100, 1)
-        })
-    repos_breakdown = sorted(repos_breakdown, key=lambda x: x["count"], reverse=True)
-
-    # 6. Process data (Insights)
-    day_counts = Counter()
-    date_set = set()
-    for c in commits:
-        day_str = c["date"].strftime("%A")
-        day_counts[day_str] += 1
-        date_set.add(c["date"].date())
-        
-    most_active_day = day_counts.most_common(1)[0][0] if day_counts else "N/A"
-    top_repo = repos_breakdown[0]["repo"] if repos_breakdown else "N/A"
-    
-    streak = 0
-    if date_set:
-        sorted_dates = sorted(list(date_set), reverse=True)
-        current_date_val = datetime.now(timezone.utc).date()
-        if sorted_dates[0] < current_date_val - timedelta(days=1):
-            streak = 0
-        else:
-            check_date = sorted_dates[0]
-            for d in sorted_dates:
-                if d == check_date:
-                    streak += 1
-                    check_date = check_date - timedelta(days=1)
-                else: break
-    
-    average_commits = round(total_commits / days, 1)
-
-    result = {
-        "commits_per_day": commits_per_day,
-        "repos_breakdown": repos_breakdown,
-        "insights": {
-            "most_active_day": most_active_day,
-            "streak": streak,
-            "top_repo": top_repo,
-            "total_summaries": total_summaries,
-            "average_commits_per_day": average_commits
-        },
-        "last_updated": datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ")
-    }
-    
-    # Store in cache
-    analytics_cache.set(cache_key, result, ttl=300)
-    
-    return result
-
-@app.get("/analytics/insights")
-async def get_insights(username: str, days: int = 30):
-    # Get total summaries from DB
-    pool = get_db_pool()
-    total_summaries = 0
-    if pool:
-        try:
-            async with pool.acquire() as conn:
-                count = await conn.fetchval('SELECT COUNT(*) FROM summaries')
-                total_summaries = count or 0
-        except Exception as e:
-            logger.error("DB error figuring out summaries count: %s", e)
-            
-    repos = await _get_user_repos(username)
-    if not repos:
-        return {"most_active_day": "N/A", "streak": 0, "top_repo": "N/A", "total_summaries": total_summaries, "average_commits_per_day": 0}
-        
-    try:
-        activity, errors = await get_activity(source="github", username=username, repos=repos, days=days)
-        commits = activity.get("commits", [])
-    except Exception as e:
-        logger.error("Failed to fetch commits: %s", e)
-        raise HTTPException(status_code=500, detail="Failed to fetch commits")
-    
-    if not commits:
-        return {"most_active_day": "N/A", "streak": 0, "top_repo": "N/A", "total_summaries": total_summaries, "average_commits_per_day": 0}
-
-    day_counts = Counter()
-    repo_counts = Counter()
-    date_set = set()
-    
-    for c in commits:
-        day_str = c["date"].strftime("%A")
-        day_counts[day_str] += 1
-        repo_counts[c["repo"]] += 1
-        date_set.add(c["date"].date())
-        
-    most_active_day = day_counts.most_common(1)[0][0] if day_counts else "N/A"
-    top_repo = repo_counts.most_common(1)[0][0] if repo_counts else "N/A"
-    
-    streak = 0
-    if date_set:
-        sorted_dates = sorted(list(date_set), reverse=True)
-        current_date_val = datetime.now(timezone.utc).date()
-        
-        if sorted_dates[0] < current_date_val - timedelta(days=1):
-            streak = 0
-        else:
-            check_date = sorted_dates[0]
-            for d in sorted_dates:
-                if d == check_date:
-                    streak += 1
-                    check_date = check_date - timedelta(days=1)
-                else:
-                    break
-    
-    average_commits = round(len(commits) / days, 1)
-    
-    return {
-        "most_active_day": most_active_day,
-        "streak": streak,
-        "top_repo": top_repo,
-        "total_summaries": total_summaries,
-        "average_commits_per_day": average_commits
-    }
 
 @app.get("/insights/metrics")
 async def get_insights_metrics(username: str, repos: str, days: int = 30):
@@ -1018,7 +367,7 @@ async def get_badge_streak(username: str):
     # In production, use standard fetching if we need actual score.
     # For now, we will query via our get_analytics_insights endpoint internally
     try:
-        from api.api import get_insights
+        from api.routers.analytics import get_insights
         res = await get_insights(username)
         streak = res.get("streak", 0)
     except Exception:
@@ -1029,10 +378,10 @@ async def get_badge_streak(username: str):
 @app.get("/badges/commits")
 async def get_badge_commits(username: str, days: int = 30):
     try:
-        from api.api import get_analytics_commits_per_day
-        from gitpulse.core.repo_reader import _get_user_repos
-        repos = await _get_user_repos(username)
+        from api.dependencies import get_user_repos
+        repos = await get_user_repos(username)
         repo_csv = ",".join(repos)
+        # get_insights_metrics is currently still in api.py
         res = await get_insights_metrics(username, repo_csv, days)
         # res returns list of {commits, prs, issues}
         total_commits = sum(d.get("commits", 0) for d in res)
@@ -1043,10 +392,10 @@ async def get_badge_commits(username: str, days: int = 30):
 @app.get("/badges/health")
 async def get_badge_health(username: str):
     try:
-        from api.api import get_insights_health
-        from gitpulse.core.repo_reader import _get_user_repos
-        repos = await _get_user_repos(username)
+        from api.dependencies import get_user_repos
+        repos = await get_user_repos(username)
         repo_csv = ",".join(repos)
+        # get_insights_health is currently still in api.py
         res = await get_insights_health(username, repo_csv)
         health_score = res.get("health_score", 0)
     except Exception:
